requirements.txt
anyio==3.7.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
Calculator==1.1.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
click==8.0.3
cmake==3.26.4
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
Django==4.1.7
django_debug_toolbar==3.8.1
duplicity==0.8.21
easyocr==1.7.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
executing==1.2.0
fake-headers==1.0.2
fasteners==0.14.1
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.2.3
fonttools==4.39.3
fqdn==1.5.1
future==0.18.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lit==16.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
MyApplication==0.1.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.8.0.74
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycairo==1.20.1
pyclipper==1.3.0.post4
pycparser==2.21
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
python-telegram-bot==20.4
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
reportlab==3.6.8
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.10.0
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.12
systemd-python==234
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.7.4
tinycss2==1.2.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wsproto==1.2.0
xdg==5
xgboost==1.7.5
xkit==0.0.0
zipp==1.0.0
emoji~=2.7.0